+++ src/ImmutableVector.php
@@ -3,9 +3,6 @@
 // Copyright (c) Lellys Informática. All rights reserved. See License.txt in the project root for license information.
 namespace Easy\Collections;
 
-use Easy\Collections\AbstractCollection;
-use Easy\Collections\ConstIndexAccessInterface;
-use Easy\Collections\ImmutableVector;
 use InvalidArgumentException;
 use OutOfBoundsException;
 use Traversable;