+++ src/Rx/ReactiveExtensionInterface.php
@@ -4,12 +4,11 @@
 namespace Easy\Collections\Rx;
 
 use Closure;
-use Easy\Collections\CollectionInterface;
 
 /**
  * Provides functionality to evaluate queries against a specific data source wherein the type of the data is not specified.
  */
-interface ReactiveExtensionInterface extends FilterInterface
+interface ReactiveExtensionInterface extends FilterableInterface, IterableInterface, AggregatableInterface
 {
 
     /**
@@ -21,12 +20,4 @@ interface ReactiveExtensionInterface extends FilterInterface
      */
     public function exists(Closure $p);
 
-    /**
-     * Iteratively reduce the collection to a single value using a callback function.
-     *
-     * @param callable $p The predicate used for reduce.
-     * @param int $initial If the optional initial is available, it will be used at the beginning of the process, or as a final result in case the array is empty.
-     * @return CollectionInterface A collection with the results of the filter operation.
-     */
-    public function reduce($p, $initial = null);
 }