+++ tests/ArrayListTest.php
@@ -259,22 +259,6 @@ class ArrayListTest extends CollectionsTestCase
         $this->assertEquals($this->coll->isEmpty(), true);
     }
 
-    public function testSlice()
-    {
-        $this->coll->addAll(array(1, 2, 3, 4));
-        $slice = $this->coll->slice(2);
-
-        $this->assertEquals(array(0 => 3, 1 => 4), $slice->toArray());
-    }
-
-//    public function testSplice()
-//    {
-//        $this->coll->addAll(array(1, 2, 3, 4));
-//        $splice = $this->coll->splice(2);
-//
-//        $this->assertEquals(array(0 => 1, 1 => 2), $splice->toArray());
-//    }
-
     public function testToMap()
     {
         $this->coll->addAll(array(1, 2, 3, 4));